More work on CMB.

+++ FinModelUtility/Fin Tests/math/BitLogicTest.cs
@@ -5,9 +5,11 @@ namespace fin.math {
   public class BitLogicTest {
     [TestMethod]
     public void ExtractFromRight() {
-      Assert.AreEqual(0b1111, BitLogic.ExtractFromRight(0b00001111, 0, 4));
-      Assert.AreEqual(0b1111, BitLogic.ExtractFromRight(0b11110000, 4, 4));
-      Assert.AreEqual(0b1011, BitLogic.ExtractFromRight(0b101100, 2, 4));
+      Assert.AreEqual((uint) 0b1111,
+                      BitLogic.ExtractFromRight(0b00001111, 0, 4));
+      Assert.AreEqual((uint) 0b1111,
+                      BitLogic.ExtractFromRight(0b11110000, 4, 4));
+      Assert.AreEqual((uint) 0b1011, BitLogic.ExtractFromRight(0b101100, 2, 4));
     }
   }
 }
